+++ oos-20190601/src/Models/StartDebugExecutionResponseBody.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Oos\V20190601\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class StartDebugExecutionResponseBody extends Model
+{
+    /**
+     * @var string
+     */
+    public $executionId;
+
+    /**
+     * @var string
+     */
+    public $requestId;
+    protected $_name = [
+        'executionId' => 'ExecutionId',
+        'requestId' => 'RequestId',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->executionId) {
+            $res['ExecutionId'] = $this->executionId;
+        }
+
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['ExecutionId'])) {
+            $model->executionId = $map['ExecutionId'];
+        }
+
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+
+        return $model;
+    }
+}