+++ oos-20190601/ChangeLog.md
@@ -1,3 +1,9 @@
+2025-12-29 Version: 2.8.0
+- Support API StartDebugExecution.
+- Update API ListGitAccounts: add request parameters BindType.
+- Update API ListGitOrganizations: add request parameters BindType.
+
+
 2025-11-07 Version: 2.7.0
 - Support API CheckGitRepoFileExists.
 - Support API CheckGitRepositoryExists.